requirements.txt
streamlit
pandas
joblib
plotly
scikit-learn
xgboost
streamlit-lottie
requests
pillow